files.size type fix, part deux

+++ repo_models/files.py
@@ -188,7 +188,7 @@ FIELDS = [
         'elasticsearch': {
             'public': True,
             'properties': {
-                'type': "integer",
+                'type': "long",
                 'store': "yes",
                 'index': "not_analyzed"
             },
